api/requirements.txt
flask==3.1.1
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.13.4
python-dotenv==1.0.0
schedule==1.2.0